Breakout catcode table code and serialization buffers
